v1.0.3: display with "_" by default, shift to copy+paste "_" version
Not sure if this behavior will be intuitive to others, but it's really
helpful for me when trying to see the magnitude of some wretchedly large
decimal value.

I also refactored the --alfred2 command line test to read from external
files instead of pasting the expected output directly in the test
script.